apps/python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
whisperx==3.1.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
yt-dlp==2023.11.16
ffmpeg-python==0.2.0
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.3
pyannote.audio==3.1.1
